15 Add alternative that uses channels
Add an alternative solution that uses channels instead of arrays. Not
sure if this is faster or not, but I wanted to try it out. I didn't find
a way to just peek into the channel to get values. You have to pop them
and then push them again if you want them to survive.

The best would have been if the first message on the channel would "fall
off" when we push a new value. And that values would stay when we peeked
them.