Solve 18.2
Easy but simple. There are probably better ways of doing it. But I'm not
a smart person so I do what is easy and readable. The summing at the end
for example, that could be done directly but it's more readable if you
do it one step at a time.