Refactor day 1
Change things around so they make more sense. Will probably use IntelliJ
for the rest of the problems. Sooooo much easier to work with than doing
it in VSCode.

Anyhow, refactored and structured so the next days can be done in the
same package. And use the input utils that I will always need.