Solve day7
I hated this. I wanted to build a tree of all the data. But that seems
to be kinda impossible in rust. How the heck can it be so hard to build
a frecking tree!? I know Rust is nice and all that when it comes to
keeping memory safe. But goddam this sucked. In the end I didn't build
the tree. I just built a map of dirs and their sizes as that was all
that was actually needed. But it is nowhere near as pretty.