Solve day 13
Well ... ehum. I did a thing. If I ever see someone push code like this
to an actual project I'm working on we're going to have a problem. Damn
this is some horrible code. I wrote it today and I'm not sure I know
what it actually does anymore. It's that bad.

The proper way is to create a enum or struct. It is either a value or a
vec. And then you implement the Order trait on that. Then you can just
pull everything in and use `cmp` to check if they are correct or not.
And it makes part 2 even less complicated.