Add solution to day 8
Man, part 2 really did not play nicely with me due to not knowing Go.
If you create an array with `arr := input[:i]` in the loop, then `arr`
will live on in the next iteration. If you do the (apparently more
correct) `var arr []string; arr = append(arr, input[:i]...)` then it all
works.
Or perhaps the first variant references the original input? I don't
know. But I do know to stay away from the first form if I want to create
temporary things in a loop.

This really messed things up and gave the an answer that were of
by the hundreds.