Solve day 15
Neato. When you write rust with actual struct implementations things get
much easier. Very much like Java actually. "Put the methods on the
object". But in rust it's "Put the function on the struct".

What I don't get is how part 1 takes 20 seconds to run and part 2 in 9
seconds. How the heck....