Solve day 14
Did part one fairly easily. Just rotated the "wrong" way for part 2
really. Rewrote part my rotation for part 2 and updated part 1 to use
that. By starting of with rotating the entire grid everything becomes a
lot easier. It makes it possible to work on one row at a time instead of
 the entire grid.